\name{cz}
\alias{cz}
\title{
C(Z)
}
\description{
Internal function.
}
\usage{
cz(w, d, dz)
}
\arguments{
  \item{w}{
the \code{dist} object of a weight matrix.
}
  \item{d}{
the \code{dist} object of a distance/dissimilarity matrix.
}
  \item{dz}{
the \code{dist} object of conditional distances.
}
}
\value{
the matrix C(Z)
}
\references{
Bui, A.T. (2024). Dimension Reduction with Prior Information for Knowledge Discovery. IEEE Transactions on Pattern Analysis and Machine Intelligence, 46, 3625-3636. https://doi.org/10.1109/TPAMI.2023.3346212
}
\author{
Anh Tuan Bui
}
