\name{cz}
\alias{cz}
\title{
C(Z)
}
\description{
Internal function.
}
\usage{
cz(w, d, dz)
}
\arguments{
  \item{w}{
the \code{dist} object of a weight matrix.
}
  \item{d}{
the \code{dist} object of a distance/dissimilarity matrix.
}
  \item{dz}{
the \code{dist} object of conditional distances.
}
}
\value{
the matrix C(Z)
}
\references{
Bui, A.T. (2021). Dimension Reduction with Prior Information for Knowledge Discovery. arXiv:2111.13646. https://arxiv.org/abs/2111.13646.
}
\author{
Anh Tuan Bui
}
